requirements.txt
flask-cors==3.0.10

Flask==1.1.2
requests_oauthlib==1.3.0

Click==7.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Werkzeug==0.16.0

sqlalchemy==1.3.23

mecab-python3==1.0.3